adds error handling on config file opening splits remote into 2 files improved functions' protection level
