+++ cacheman.go
@@ -14,14 +14,12 @@ var Cfg Config
 
 func main() {
 	local.GetConfig(&Cfg)
-	local.GetMirrorList(&Cfg)
-	http.HandleFunc("/", HandleReq)
+	http.HandleFunc("/", handleReq)
 	log.Fatal(http.ListenAndServe(":8080", nil))
 }
 
-func HandleReq(w http.ResponseWriter, r *http.Request) {
+func handleReq(w http.ResponseWriter, r *http.Request) {
 	NowStr := time.Now().Format(time.Kitchen)
-	//TODO: review public and private functions
 	RequestedLocalPath := Cfg.CacheDir + "/" + r.URL.Path[1:] //add cachedir to path, to not check in /
 	RequestedPath := r.URL.Path[1:]
 	fmt.Printf("[SERVER %s] File requested: %s\n", NowStr, RequestedPath)
